Add files via upload
Here, we are randomly generating the birth date and birth month of unknown persons and inserting it into the list. After the insertion, we are checking if any two or more persons are having the same birth date and birth month. If they have any common birthdays then we are displaying it and if they do not then we are displaying the other.